ssl function is moved to main.go
